Move redux store setup to its own file
